Move MagPoseFactor to gtsam
