+++ gtsam/navigation/MagPoseFactor.h
